requirements.txt
requests
pandas
python-dotenv
markdown2
reportlab
